+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.2.5] - 20210716.
+
+* bump polkawallet_sdk 0.2.5.
+* update earn pages for Karura.
+* update swap for Karura.
+* add Karura dex provisioning.
+
 ## [0.2.2] - 20210623.
 
 * bump polkawallet_sdk 0.2.2.